Solicitação ao Tutor. -Efetuei correções. Esta apresentando erros: -Aparece sinal de Loadingm, porem apois não aparece nem listagem nem Mensagem de Erro(Ops, aconteceu um erro. Tente novamente mais tarde.) Não esta retornando listagem. Obrigado.

+++ Dev_Android_Kolin_Python/Modulo17_Exercicio_005/app/src/main/java/com/example/filmes/data/Filmes.kt
@@ -8,7 +8,7 @@ class Filmes (
     val id: Int?,
     val title: String?,
     val overview_: String?,
-    val popularity: Int?,
+    val popularity: Double?,
     val poster_path: String
 ){
